--- server/src/acuity/functions/trpc/acuity.ts
@@ -1,4 +1,0 @@
-import { acuityRouter } from '../../core/acuity-router'
-import { onRequestTrpc } from '../../../trpc/trpc.adapter'
-
-export const acuity = onRequestTrpc(acuityRouter)
