--- server/src/events/functions/trpc/events.ts
@@ -1,4 +1,0 @@
-import { onRequestTrpc } from '../../../trpc/trpc.adapter'
-import { eventsRouter } from '../../core/events-router'
-
-export const events = onRequestTrpc(eventsRouter)
