+++ server/src/holidayPrograms/core/schedule-holiday-program.ts
