+++ server/src/holidayPrograms/core/send-confirmation-email.ts
