+++ server/src/partyBookings/functions/pubsub/handlePartyFormSubmission.ts
@@ -5,7 +5,7 @@ import { DateTime } from 'luxon'
 import { DatabaseClient } from '../../../firebase/DatabaseClient'
 import { MailClient } from '../../../sendgrid/MailClient'
 import { logError, onMessagePublished, throwFunctionsError } from '../../../utilities'
-import { FormMapper } from '../../core/FormMapper'
+import { FormMapper } from '../../core/form-mapper'
 import { getBookingAdditions, getBookingCreations } from '../../core/utils'
 
 export const handlePartyFormSubmission = onMessagePublished('handlePartyFormSubmission', async (responses) => {
