--- server/src/partyBookings/functions/trpc/parties.ts
@@ -1,4 +1,0 @@
-import { onRequestTrpc } from '../../../trpc/trpc.adapter'
-import { partiesRouter } from '../../core/parties-router'
-
-export const parties = onRequestTrpc(partiesRouter)
