+++ server/src/trpc/trpc.app-router.ts
@@ -1,12 +1,14 @@
-import { eventsRouter } from '../events/core/events-router'
-import { partiesRouter } from '../partyBookings/core/parties-router'
+import { acuityRouter } from '../acuity/functions/trpc/trpc.acuity'
+import { eventsRouter } from '../events/functions/trpc/trpc.events'
+import { holidayProgramsRouter } from '../holidayPrograms/functions/trpc/trpc.holiday-programs'
+import { partiesRouter } from '../partyBookings/functions/trpc/trpc.parties'
 import { router } from './trpc'
-import { acuityRouter } from '../acuity/core/acuity-router'
 
 const appRouter = router({
     parties: partiesRouter,
     events: eventsRouter,
     acuity: acuityRouter,
+    holidayPrograms: holidayProgramsRouter,
 })
 
 export type AppRouter = typeof appRouter
